requirements.txt
tensorflow
torch
numpy
pandas
tqdm
matplotlib
python-dotenv